refactor: MemberList interface 파일분리
