+++ src/model/page.ts
@@ -1,0 +1,10 @@
+import { Member } from './member';
+
+export interface MemberList {
+  content: Member[];
+  hasNext: boolean;
+  currentPage: number;
+  totalElements: number;
+  totalPage: number;
+  pageSize: number;
+}
